requirements.txt
# Core dependencies
pikepdf>=8.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=0.6.21
Pillow>=10.0.0
mutagen>=1.47.0
PyExifTool>=0.5.5
pymediainfo>=6.0.1

# Office document processing
odfpy>=1.4.1
striprtf>=0.0.26

# Web GUI
Flask>=3.0.0
Flask-CORS>=4.0.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
python-magic-bin>=0.4.14; platform_system == "Windows"
python-magic>=0.4.27; platform_system != "Windows"

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
